testTmp/requirements.txt
streamlit
pandas
numpy
python-dotenv
openai